requirements.txt
sqlmodel
fastapi[standard]
pytest
httpx
pyjwt
bcrypt